Rebuilt the website, added responsive design
It was proving too difficult to work on the site as it was standing, so I just rebuilt the entire site from scratch again. This was to help me restructure the code in a way that made more sense to me, and it allowed me to add the functionality to the site that I wanted without deleting code I'd already written. There are a few bugs, like the <section> tags are refusing to space properly, and the navigation links don't seem to want to make the page jump down to the <section> that they are supposed to link to.